+++ ok-workout-be-app-ktor/src/main/kotlin/ru/otus/otuskotlin/workout/plugins/Routing.kt
@@ -1,12 +1,12 @@
-package ru.otus.plugins
+package ru.otus.otuskotlin.workout.plugins
 
 import ExerciseService
 import WorkoutService
 import io.ktor.application.*
 import io.ktor.response.*
 import io.ktor.routing.*
-import ru.otus.exercise
-import ru.otus.workout
+import ru.otus.otuskotlin.workout.exercise
+import ru.otus.otuskotlin.workout.workout
 
 fun Application.configRouting() {
